只需要把权限套用即可。 一周半的时间，每天10个小时以上，jbpm流程的最佳实践。over

+++ src/org/supermy/workflow/action/FinanceProcessAction.java
@@ -21,7 +21,7 @@ public class FinanceProcessAction implements ActionHandler {
 		String actorId = executionContext.getJbpmContext().getActorId();
 		log.debug("actorId:{}", actorId);
 
-		int count = (Integer) executionContext.getProcessInstance()
+		double count = (Double) executionContext.getProcessInstance()
 				.getContextInstance().getVariable("money");
 		String mainID = (String) executionContext.getProcessInstance()
 				.getContextInstance().getVariable("businessKey");
